feat(ship): grid is clickable and interacts with API
