+++ package.json
@@ -18,6 +18,7 @@
     "axios": "^1.4.0",
     "pinia": "^2.0.36",
     "sass": "^1.62.1",
+    "socket.io-client": "^4.7.2",
     "vue": "^3.3.2",
     "vue-router": "^4.2.0"
   },
